+++ packages/client/components/promptResponse/tiptapConfig.ts
@@ -1,6 +1,24 @@
+import {Editor} from '@tiptap/core'
 import Link from '@tiptap/extension-link'
 import Placeholder from '@tiptap/extension-placeholder'
 import StarterKit from '@tiptap/starter-kit'
+
+export const getLinkProps = (editor: Editor) => {
+  const href: string | undefined = editor.getAttributes('link').href
+  let {from, to} = editor.view.state.selection
+  if (to === from && editor.isActive('link')) {
+    // If the cursor is on a link, but hasn't selected any specific part of the link text, expand the selection
+    // to the full link.
+    editor.commands.setTextSelection({to: to - 1, from: from - 1})
+    editor.commands.extendMarkRange('link')
+    const selection = editor.view.state.selection
+    to = selection.to
+    from = selection.from
+  }
+  const text = editor.state.doc.textBetween(from, to, '')
+
+  return {text, href: href ?? ''}
+}
 
 /**
  * Returns tip tap extensions configuration shared by the client and the server
@@ -9,7 +27,7 @@ import StarterKit from '@tiptap/starter-kit'
  */
 export const createEditorExtensions = (
   isReadOnly?: boolean,
-  onOpenLinkMenu?,
+  setLinkMenuProps?,
   setSelectedHref?,
   placeholder?: string
 ) => [
@@ -17,23 +35,12 @@ export const createEditorExtensions = (
   Link.extend({
     addKeyboardShortcuts() {
       return {
-        // ↓ your new keyboard shortcut
         'Mod-k': () => {
-          if (!onOpenLinkMenu) {
+          if (!setLinkMenuProps) {
             return false
           }
 
-          const href: string | undefined = this.editor.getAttributes('link').href
-          let {from, to} = this.editor.view.state.selection
-          if (to === from) {
-            this.editor.commands.setTextSelection({to: to - 1, from: from - 1})
-            this.editor.commands.extendMarkRange('link')
-            const selection = this.editor.view.state.selection
-            to = selection.to
-            from = selection.from
-          }
-          const text = this.editor.state.doc.textBetween(from, to, '')
-          onOpenLinkMenu({text, href})
+          setLinkMenuProps(getLinkProps(this.editor))
           return true
         }
       }
